requirements.txt

boto3==1.14.44
botocore==1.17.44
dj-database-url==0.5.0
Django==3.0.5
django-heroku==0.3.1
django-storages==1.9.1
django-tagging==0.5.0
gunicorn==20.0.4
Pillow==7.2.0
psycopg2==2.8.5
psycopg2-binary==2.8.4
python-dateutil==2.8.1
python-decouple==3.3
s3transfer==0.3.3
shellescape==3.8.1
simplejson==3.17.2
sqlparse==0.3.1
stripe==2.54.0
whitenoise==5.2.0